Revert "updated and merge branch."
This reverts commit 4f1dfe8036ad97ba7522f014ee8017ccf6b6bd12.

+++ ams-setup.sql
@@ -55,7 +55,7 @@ create table Customer
 	PRIMARY KEY (cid));
 
 create table Purchase
-	(receiptId number(15), 
+	(receiptId integer, 
 	pdate DATE default(sysdate),
 	cid number(10), 
 	cardnum varchar(16),
